feat: added checking if user exist

+++ src/users/users.service.ts
@@ -28,11 +28,19 @@ export class UsersService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
+    const isExist = await this.userRepository.findOne({ where: { id: id } });
+    if (!isExist) {
+      return null;
+    }
     await this.userRepository.update(id, updateUserDto);
     return this.userRepository.findOne({ where: { id: id } });
   }
 
   async remove(id: number) {
+    const isExist = await this.userRepository.findOne({ where: { id: id } });
+    if (!isExist) {
+      return;
+    }
     this.userRepository.delete(id);
   }
 }
